Add env variable to determine oauth strategy
