+++ app/services/auth/index.js
@@ -1,8 +1,10 @@
 import setupOAuth from './oauth';
 import { setupJwtAuth } from './jwt';
 
+const config = require('config');
+
 async function setupAuth (server) {
-  await setupOAuth(server);
+  await setupOAuth(server, config.get('oauthStrategy'));
   await setupJwtAuth(server);
 }
